# api/openapi.json
{
  "components": {
    "schemas": {
      "AbsResult": {
        "properties": {
          "ip_approval_needed": {
            "title": "Ip Approval Needed",
            "type": "boolean"
          },
          "route": {
            "enum": [
              "nba_approval",
              "sbb_intimation",
              "exempt_check",
              "not_applicable"
            ],
            "title": "Route",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Line"
            },
            "title": "Steps",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "verified_on": {
            "default": "2026-09-17",
            "title": "Verified On",
            "type": "string"
          }
        },
        "required": [
          "route",
          "summary",
          "steps",
          "ip_approval_needed"
        ],
        "title": "AbsResult",
        "type": "object"
      },
      "AbsStep": {
        "properties": {
          "next_question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Question"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbsResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AbsStep",
        "type": "object"
      },
...
      "AskRequest": {
        "properties": {
          "context": {
            "items": {
              "$ref": "#/components/schemas/ContextItem"
            },
            "maxItems": 10,
            "title": "Context",
            "type": "array"
          },
          "jurisdiction_mode": {
            "default": "BOTH",
            "enum": [
              "IN",
              "INTL",
              "BOTH"
            ],
            "title": "Jurisdiction Mode",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "persona": {
            "anyOf": [
              {
                "enum": [
                  "practitioner",
                  "researcher",
                  "startup",
                  "cultivator"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
...
      "ClassifyRequest": {
        "properties": {
          "answers": {
            "additionalProperties": {
              "type": "boolean"
            },
            "default": {},
            "title": "Answers",
            "type": "object"
          }
        },
        "title": "ClassifyRequest",
        "type": "object"
      },
      "ClassifyStep": {
        "properties": {
          "asked": {
            "title": "Asked",
            "type": "integer"
          },
          "next_question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Question"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassificationResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "asked"
        ],
        "title": "ClassifyStep",
        "type": "object"
      },
...
      "ConsentRequest": {
        "properties": {
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "scope": {
            "enum": [
              "assistant",
              "transcript"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "granted"
        ],
        "title": "ConsentRequest",
        "type": "object"
      },
      "ConsentState": {
        "properties": {
          "assistant": {
            "title": "Assistant",
            "type": "boolean"
          },
          "transcript": {
            "title": "Transcript",
            "type": "boolean"
          }
        },
        "required": [
          "assistant",
          "transcript"
        ],
        "title": "ConsentState",
        "type": "object"
      },
...
      "EscalateRequest": {
        "properties": {
          "answer_id": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Id"
          },
          "contact": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "message": {
            "default": "",
            "maxLength": 2000,
            "title": "Message",
            "type": "string"
          }
        },
        "title": "EscalateRequest",
        "type": "object"
      },
      "EscalateResult": {
        "properties": {
          "facilitators": {
            "items": {
              "$ref": "#/components/schemas/FacilitatorOut"
            },
            "title": "Facilitators",
            "type": "array"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "string"
          }
        },
        "required": [
          "ticket_id",
          "status",
          "facilitators",
          "note"
        ],
        "title": "EscalateResult",
        "type": "object"
      },
...
      "GI": {
        "properties": {
          "tags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GITag"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "title": "GI",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MaterialIPProfile": {
        "description": "`cite` fields hold a manifest source id, or 'unknown' when unverified.",
        "properties": {
          "biodiversity": {
            "$ref": "#/components/schemas/Biodiversity"
          },
          "deposit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Deposit"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "drugSchedules": {
            "$ref": "#/components/schemas/DrugSchedules"
          },
          "export": {
            "$ref": "#/components/schemas/Export"
          },
          "gi": {
            "$ref": "#/components/schemas/GI"
          },
          "kind": {
            "enum": [
              "plant",
              "microbe",
              "animal",
              "mineral"
            ],
            "title": "Kind",
            "type": "string"
          },
          "lastVerified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "ISO date; null = never verified",
            "title": "Lastverified"
          },
          "monographs": {
            "$ref": "#/components/schemas/Monographs"
          },
          "patentability": {
            "$ref": "#/components/schemas/Patentability"
          },
          "patents": {
            "$ref": "#/components/schemas/Patents"
          },
          "tk": {
            "$ref": "#/components/schemas/TraditionalKnowledge"
          },
          "wildlife": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Wildlife"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "kind",
          "tk",
          "patentability",
          "patents",
          "gi",
          "biodiversity",
          "export",
          "drugSchedules",
          "monographs"
        ],
        "title": "MaterialIPProfile",
        "type": "object"
      },
...
      "PurgeResult": {
        "properties": {
          "deleted": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Deleted",
            "type": "object"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "PurgeResult",
        "type": "object"
      },
      "Question": {
        "properties": {
          "help": {
            "default": "",
            "title": "Help",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text"
        ],
        "title": "Question",
        "type": "object"
      },
      "RegistryOut": {
        "properties": {
          "cite_locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cite Locator"
          },
          "cite_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cite Source Id"
          },
          "fee_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee Note"
          },
          "forms": {
            "items": {},
            "title": "Forms",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "jurisdiction": {
            "title": "Jurisdiction",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "regime": {
            "items": {
              "type": "string"
            },
            "title": "Regime",
            "type": "array"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "jurisdiction",
          "regime",
          "url",
          "forms",
          "fee_note",
          "cite_locator",
          "cite_source_id"
        ],
        "title": "RegistryOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "IP-SAKTI Sahayak API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/abs": {
      "post": {
        "operationId": "abs_helper_abs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbsStep"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abs Helper",
        "tags": [
          "assistant"
        ]
      }
    },
    "/ask": {
      "post": {
        "operationId": "ask_ask_post",
        "parameters": [
          {
            "description": "anonymous id from the browser",
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "description": "anonymous id from the browser",
              "title": "X-Session-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "status, answer, done"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "assistant"
        ]
      }
    },
    "/classify": {
      "post": {
        "description": "One question at a time: send the answers so far, get the next question or the result.",
        "operationId": "classify_formulation_classify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifyStep"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Classify Formulation",
        "tags": [
          "assistant"
        ]
      }
    },
    "/consent": {
      "get": {
        "operationId": "get_consent_consent_get",
        "parameters": [
          {
            "description": "anonymous id from the browser",
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "description": "anonymous id from the browser",
              "title": "X-Session-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Consent",
        "tags": [
          "privacy"
        ]
      },
      "post": {
        "operationId": "set_consent_consent_post",
        "parameters": [
          {
            "description": "anonymous id from the browser",
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "description": "anonymous id from the browser",
              "title": "X-Session-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Consent",
        "tags": [
          "privacy"
        ]
      }
    },
    "/escalate": {
      "post": {
        "operationId": "escalate_escalate_post",
        "parameters": [
          {
            "description": "anonymous id from the browser",
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "description": "anonymous id from the browser",
              "title": "X-Session-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Escalate",
        "tags": [
          "assistant"
        ]
      }
    },
...
    "/materials/{kind}/{material_id}/ipr": {
      "get": {
        "operationId": "material_ipr_materials__kind___material_id__ipr_get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "plant",
                "microbe",
                "animal",
                "mineral"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialIPProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Material Ipr",
        "tags": [
          "corpus"
        ]
      }
    },
    "/me": {
      "delete": {
        "operationId": "delete_me_me_delete",
        "parameters": [
          {
            "description": "anonymous id from the browser",
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "description": "anonymous id from the browser",
              "title": "X-Session-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Me",
        "tags": [
          "privacy"
        ]
      }
    },
    "/registry": {
      "get": {
        "operationId": "list_registries_registry_get",
        "parameters": [
          {
            "in": "query",
            "name": "regime",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Regime"
            }
          },
          {
            "in": "query",
            "name": "jurisdiction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "IN",
                    "INTL"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jurisdiction"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RegistryOut"
                  },
                  "title": "Response List Registries Registry Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Registries",
        "tags": [
          "corpus"
        ]
      }
    },
...
  }
}